Add styles to header and button
The name of some classes such as `left-links` need to be updated in a
separate commit. The naming does't make sense in the `left-links`
example since it'll only be one item and the item may not even have a
link attached in the final product.

The `links` class visually looks correct until viewed on a large screen
or when the webpage is zoomed out, unlike `hero-content` which stays a
fixed width until it has to collaspe on a small screen or zoomed in. A
possible fix would be to encaspulate `links` into another container but
the worry here is adding complexity by adding container inside of
container.